requirements.txt
apache-airflow-providers-postgres==5.6.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0